pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-Customer_Support_Gym_2"
version = "0.1.0"
description = "Customer Support Gym 2 environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.2",
    # .env file loading — reads API keys without hardcoding them
    "python-dotenv>=1.0.0",
    # OpenAI-compatible client used by inference.py
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m Customer_Support_Gym_2.server.app
server = "Customer_Support_Gym_2.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["Customer_Support_Gym_2", "Customer_Support_Gym_2.server"]
package-dir = { "Customer_Support_Gym_2" = ".", "Customer_Support_Gym_2.server" = "server" }
[tool.pyright]
include = ["."]
extraPaths = ["."]
executionEnvironments = [
    { root = "." }
]
